patch: move all get_rgba8 calls as impls for PicData trait
