+++ src/pic/pic.rs
@@ -4,20 +4,17 @@ use std::sync::Arc;
 
 use crate::parser::{aptnode::APTNode, lexer::Lexer, token::Token};
 use crate::pic::actual_picture::ActualPicture;
-use crate::pic::coordinatesystem::{cartesian_to_polar, CoordinateSystem};
-use crate::pic::data::gradient::{GradientData, GRADIENT_SIZE};
+use crate::pic::coordinatesystem::CoordinateSystem;
+use crate::pic::data::gradient::GradientData;
 use crate::pic::data::grayscale::GrayscaleData;
 use crate::pic::data::hsv::HSVData;
 use crate::pic::data::mono::MonoData;
 use crate::pic::data::rgb::RGBData;
 use crate::pic::data::PicData;
-use crate::pic::ggez_utility::lerp_color;
-use crate::vm::stackmachine::StackMachine;
 
 use ggez::graphics::Color;
 use rand::prelude::*;
 use rand::rngs::StdRng;
-use rayon::prelude::*;
 use simdeez::avx2::*;
 use simdeez::scalar::*;
 use simdeez::sse2::*;
@@ -164,424 +161,11 @@ impl Pic {
         t: f32,
     ) -> Vec<u8> {
         match self {
-            Pic::Mono(data) => Pic::get_rgba8_mono::<S>(data, threaded, pics, w, h, t),
-            Pic::Grayscale(data) => Pic::get_rgba8_grayscale::<S>(data, threaded, pics, w, h, t),
-            Pic::Gradient(data) => Pic::get_rgba8_gradient::<S>(data, threaded, pics, w, h, t),
-            Pic::RGB(data) => Pic::get_rgba8_rgb::<S>(data, threaded, pics, w, h, t),
-            Pic::HSV(data) => Pic::get_rgba8_hsv::<S>(data, threaded, pics, w, h, t),
-        }
-    }
-
-    fn get_rgba8_gradient<S: Simd>(
-        data: &GradientData,
-        threaded: bool,
-        pics: Arc<HashMap<String, ActualPicture>>,
-        w: usize,
-        h: usize,
-        t: f32,
-    ) -> Vec<u8> {
-        unsafe {
-            let ts = S::set1_ps(t);
-            let wf = S::set1_ps(w as f32);
-            let hf = S::set1_ps(h as f32);
-            let vec_len = w * h * 4;
-            let mut result = Vec::<u8>::with_capacity(vec_len);
-            result.set_len(vec_len);
-            let sm = StackMachine::<S>::build(&data.index);
-            /*
-            let mut min = 999999.0;
-            let mut max = -99999.0;
-            */
-
-            let color_count = data.colors.iter().filter(|(_, stop)| !stop).count();
-            let mut gradient = Vec::<Color>::new(); //todo actually compute this
-            let step = (GRADIENT_SIZE as f32 / color_count as f32) / GRADIENT_SIZE as f32;
-            let mut positions = Vec::<f32>::new();
-            positions.push(0.0);
-            let mut pos = step;
-            for i in 1..data.colors.len() - 1 {
-                let (_, stop) = data.colors[i];
-                if stop {
-                    positions.push(*positions.last().unwrap());
-                } else {
-                    positions.push(pos);
-                    pos += step;
-                }
-            }
-            positions.push(1.0);
-
-            for i in 0..GRADIENT_SIZE {
-                let pct = i as f32 / GRADIENT_SIZE as f32;
-                let color2pos = positions.iter().position(|n| *n >= pct).unwrap();
-                if color2pos == 0 {
-                    gradient.push(data.colors[0].0);
-                } else {
-                    let color1 = data.colors[color2pos - 1].0;
-                    let color2 = data.colors[color2pos].0;
-                    let pct2 = positions[color2pos];
-                    let pct1 = positions[color2pos - 1];
-                    let range = pct2 - pct1;
-                    let pct = (pct - pct1) / range;
-                    gradient.push(lerp_color(color1, color2, pct));
-                }
-            }
-
-            let process = |(y_pixel, chunk): (usize, &mut [u8])| {
-                let mut stack = Vec::with_capacity(sm.instructions.len());
-                stack.set_len(sm.instructions.len());
-
-                let y = S::set1_ps((y_pixel as f32 / h as f32) * 2.0 - 1.0);
-                let x_step = 2.0 / (w - 1) as f32;
-                let mut x = S::setzero_ps();
-                for i in (0..S::VF32_WIDTH).rev() {
-                    x[i] = -1.0 + (x_step * i as f32);
-                }
-                let x_step = S::set1_ps(x_step * S::VF32_WIDTH as f32);
-
-                for i in (0..w * 4).step_by(S::VF32_WIDTH * 4) {
-                    let v = if data.coord == CoordinateSystem::Cartesian {
-                        sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                    } else {
-                        let (r, theta) = cartesian_to_polar::<S>(x, y);
-                        sm.execute(&mut stack, pics.clone(), r, theta, ts, wf, hf)
-                    };
-                    let scaled_v = (v + S::set1_ps(1.0)) * S::set1_ps(0.5);
-                    let index = S::cvtps_epi32(scaled_v * S::set1_ps(GRADIENT_SIZE as f32));
-
-                    for j in 0..S::VF32_WIDTH {
-                        let c = gradient[index[j] as usize % GRADIENT_SIZE];
-                        let j4 = j * 4;
-                        chunk[i + j4] = (c.r * 255.0) as u8;
-                        chunk[i + 1 + j4] = (c.g * 255.0) as u8;
-                        chunk[i + 2 + j4] = (c.b * 255.0) as u8;
-                        chunk[i + 3 + j4] = 255 as u8;
-                    }
-                    x = x + x_step;
-                }
-            };
-
-            if threaded {
-                result.par_chunks_mut(4 * w).enumerate().for_each(process);
-            } else {
-                result.chunks_exact_mut(4 * w).enumerate().for_each(process);
-            }
-
-            // println!("min:{} max:{} range:{}",min,max,max-min);
-            result
-        }
-    }
-
-    fn get_rgba8_grayscale<S: Simd>(
-        data: &GrayscaleData,
-        threaded: bool,
-        pics: Arc<HashMap<String, ActualPicture>>,
-        w: usize,
-        h: usize,
-        t: f32,
-    ) -> Vec<u8> {
-        unsafe {
-            let ts = S::set1_ps(t);
-            let wf = S::set1_ps(w as f32);
-            let hf = S::set1_ps(h as f32);
-            let vec_len = w * h * 4;
-            let mut result = Vec::<u8>::with_capacity(vec_len);
-            result.set_len(vec_len);
-            let sm = StackMachine::<S>::build(&data.c);
-            /*
-            let mut min = 999999.0;
-            let mut max = -99999.0;
-            */
-
-            let process = |(y_pixel, chunk): (usize, &mut [u8])| {
-                let mut stack = Vec::with_capacity(sm.instructions.len());
-                stack.set_len(sm.instructions.len());
-
-                let y = S::set1_ps((y_pixel as f32 / h as f32) * 2.0 - 1.0);
-                let x_step = 2.0 / (w - 1) as f32;
-                let mut x = S::setzero_ps();
-                for i in (0..S::VF32_WIDTH).rev() {
-                    x[i] = -1.0 + (x_step * i as f32);
-                }
-                let x_step = S::set1_ps(x_step * S::VF32_WIDTH as f32);
-
-                for i in (0..w * 4).step_by(S::VF32_WIDTH * 4) {
-                    let v = if data.coord == CoordinateSystem::Cartesian {
-                        sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                    } else {
-                        let (r, theta) = cartesian_to_polar::<S>(x, y);
-                        sm.execute(&mut stack, pics.clone(), r, theta, ts, wf, hf)
-                    };
-
-                    // if v[0] > max { max = v[0]; }
-                    // if v[0] < min { min = v[0]; }
-
-                    let cs = (v + S::set1_ps(1.0)) * S::set1_ps(127.5);
-
-                    for j in 0..S::VF32_WIDTH {
-                        let c = (cs[j] as i32 % 256) as u8;
-                        let j4 = j * 4;
-                        chunk[i + j4] = c;
-                        chunk[i + 1 + j4] = c;
-                        chunk[i + 2 + j4] = c;
-                        chunk[i + 3 + j4] = 255 as u8;
-                    }
-                    x = x + x_step;
-                }
-            };
-
-            if threaded {
-                result.par_chunks_mut(4 * w).enumerate().for_each(process);
-            } else {
-                result.chunks_exact_mut(4 * w).enumerate().for_each(process);
-            }
-            // println!("min:{} max:{} range:{}",min, max, max-min);
-            result
-        }
-    }
-
-    fn get_rgba8_mono<S: Simd>(
-        data: &MonoData,
-        threaded: bool,
-        pics: Arc<HashMap<String, ActualPicture>>,
-        w: usize,
-        h: usize,
-        t: f32,
-    ) -> Vec<u8> {
-        unsafe {
-            let ts = S::set1_ps(t);
-            let wf = S::set1_ps(w as f32);
-            let hf = S::set1_ps(h as f32);
-            let vec_len = w * h * 4;
-            let mut result = Vec::<u8>::with_capacity(vec_len);
-            result.set_len(vec_len);
-            let sm = StackMachine::<S>::build(&data.c);
-            /*
-            let mut min = 999999.0;
-            let mut max = -99999.0;
-            */
-
-            let process = |(y_pixel, chunk): (usize, &mut [u8])| {
-                let mut stack = Vec::with_capacity(sm.instructions.len());
-                stack.set_len(sm.instructions.len());
-
-                let y = S::set1_ps((y_pixel as f32 / h as f32) * 2.0 - 1.0);
-                let x_step = 2.0 / (w - 1) as f32;
-                let mut x = S::setzero_ps();
-                for i in (0..S::VF32_WIDTH).rev() {
-                    x[i] = -1.0 + (x_step * i as f32);
-                }
-                let x_step = S::set1_ps(x_step * S::VF32_WIDTH as f32);
-
-                for i in (0..w * 4).step_by(S::VF32_WIDTH * 4) {
-                    let v = if data.coord == CoordinateSystem::Cartesian {
-                        sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                    } else {
-                        let (r, theta) = cartesian_to_polar::<S>(x, y);
-                        sm.execute(&mut stack, pics.clone(), r, theta, ts, wf, hf)
-                    };
-
-                    for j in 0..S::VF32_WIDTH {
-                        let c = if v[j] >= 0.0 { 255 } else { 0 };
-                        let j4 = j * 4;
-                        chunk[i + j4] = c;
-                        chunk[i + 1 + j4] = c;
-                        chunk[i + 2 + j4] = c;
-                        chunk[i + 3 + j4] = 255 as u8;
-                    }
-                    x = x + x_step;
-                }
-            };
-
-            if threaded {
-                result.par_chunks_mut(4 * w).enumerate().for_each(process);
-            } else {
-                result.chunks_exact_mut(4 * w).enumerate().for_each(process);
-            }
-            // println!("min:{} max:{} range:{}",min,max,max-min);
-            result
-        }
-    }
-
-    fn get_rgba8_rgb<S: Simd>(
-        data: &RGBData,
-        threaded: bool,
-        pics: Arc<HashMap<String, ActualPicture>>,
-        w: usize,
-        h: usize,
-        t: f32,
-    ) -> Vec<u8> {
-        unsafe {
-            let ts = S::set1_ps(t);
-            let wf = S::set1_ps(w as f32);
-            let hf = S::set1_ps(h as f32);
-
-            let vec_len = w * h * 4;
-            let mut result = Vec::<u8>::with_capacity(vec_len);
-            result.set_len(vec_len);
-
-            let r_sm = StackMachine::<S>::build(&data.r);
-            let g_sm = StackMachine::<S>::build(&data.g);
-            let b_sm = StackMachine::<S>::build(&data.b);
-            let max_len = *[
-                r_sm.instructions.len(),
-                g_sm.instructions.len(),
-                b_sm.instructions.len(),
-            ]
-            .iter()
-            .max()
-            .unwrap();
-
-            let process = |(y_pixel, chunk): (usize, &mut [u8])| {
-                let mut stack = Vec::with_capacity(max_len);
-                stack.set_len(max_len);
-                let y = S::set1_ps((y_pixel as f32 / h as f32) * 2.0 - 1.0);
-                let x_step = 2.0 / (w - 1) as f32;
-                let mut x = S::setzero_ps();
-                for i in (0..S::VF32_WIDTH).rev() {
-                    x[i] = -1.0 + (x_step * i as f32);
-                }
-                let x_step = S::set1_ps(x_step * S::VF32_WIDTH as f32);
-
-                for i in (0..w * 4).step_by(S::VF32_WIDTH * 4) {
-                    let (rs, gs, bs) = if data.coord == CoordinateSystem::Cartesian {
-                        let rs = (r_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(128.0);
-                        let gs = (g_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(128.0);
-                        let bs = (b_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(128.0);
-                        (rs, gs, bs)
-                    } else {
-                        let (x, y) = cartesian_to_polar::<S>(x, y);
-                        let rs = (r_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(128.0);
-                        let gs = (g_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(128.0);
-                        let bs = (b_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(128.0);
-                        (rs, gs, bs)
-                    };
-
-                    for j in 0..S::VF32_WIDTH {
-                        let r = (rs[j] as i32 % 255) as u8;
-                        let g = (gs[j] as i32 % 255) as u8;
-                        let b = (bs[j] as i32 % 255) as u8;
-                        let j4 = j * 4;
-                        chunk[i + j4] = r;
-                        chunk[i + 1 + j4] = g;
-                        chunk[i + 2 + j4] = b;
-                        chunk[i + 3 + j4] = 255 as u8;
-                    }
-                    x = x + x_step;
-                }
-            };
-            if threaded {
-                result.par_chunks_mut(4 * w).enumerate().for_each(process);
-            } else {
-                result.chunks_exact_mut(4 * w).enumerate().for_each(process);
-            }
-
-            result
-        }
-    }
-
-    fn get_rgba8_hsv<S: Simd>(
-        data: &HSVData,
-        threaded: bool,
-        pics: Arc<HashMap<String, ActualPicture>>,
-        w: usize,
-        h: usize,
-        t: f32,
-    ) -> Vec<u8> {
-        unsafe {
-            let ts = S::set1_ps(t);
-            let wf = S::set1_ps(w as f32);
-            let hf = S::set1_ps(h as f32);
-            let vec_len = w * h * 4;
-            let mut result = Vec::<u8>::with_capacity(vec_len);
-            result.set_len(vec_len);
-
-            let h_sm = StackMachine::<S>::build(&data.h);
-            let s_sm = StackMachine::<S>::build(&data.s);
-            let v_sm = StackMachine::<S>::build(&data.v);
-            let max_len = *[
-                h_sm.instructions.len(),
-                s_sm.instructions.len(),
-                v_sm.instructions.len(),
-            ]
-            .iter()
-            .max()
-            .unwrap();
-
-            let process = |(y_pixel, chunk): (usize, &mut [u8])| {
-                let mut stack = Vec::with_capacity(max_len);
-                stack.set_len(max_len);
-                let y = S::set1_ps((y_pixel as f32 / h as f32) * 2.0 - 1.0);
-                let x_step = 2.0 / (w - 1) as f32;
-                let mut x = S::setzero_ps();
-                for i in (0..S::VF32_WIDTH).rev() {
-                    x[i] = -1.0 + (x_step * i as f32);
-                }
-                let x_step = S::set1_ps(x_step * S::VF32_WIDTH as f32);
-
-                for i in (0..w * 4).step_by(S::VF32_WIDTH * 4) {
-                    let (hs, ss, vs) = if data.coord == CoordinateSystem::Cartesian {
-                        let hs = (h_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(0.5);
-                        let ss = (s_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(0.5);
-                        let vs = (v_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(0.5);
-                        (hs, ss, vs)
-                    } else {
-                        let (x, y) = cartesian_to_polar::<S>(x, y);
-                        let hs = (h_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(0.5);
-                        let ss = (s_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(0.5);
-                        let vs = (v_sm.execute(&mut stack, pics.clone(), x, y, ts, wf, hf)
-                            + S::set1_ps(1.0))
-                            * S::set1_ps(0.5);
-                        (hs, ss, vs)
-                    };
-
-                    let (mut rs, mut gs, mut bs) =
-                        hsv_to_rgb::<S>(wrap_0_1::<S>(hs), wrap_0_1::<S>(ss), wrap_0_1::<S>(vs));
-                    rs = rs * S::set1_ps(255.0);
-                    gs = gs * S::set1_ps(255.0);
-                    bs = bs * S::set1_ps(255.0);
-                    for j in 0..S::VF32_WIDTH {
-                        let r = (rs[j] as i32 % 255) as u8;
-                        let g = (gs[j] as i32 % 255) as u8;
-                        let b = (bs[j] as i32 % 255) as u8;
-                        let j4 = j * 4;
-                        chunk[i + j4] = r;
-                        chunk[i + 1 + j4] = g;
-                        chunk[i + 2 + j4] = b;
-                        chunk[i + 3 + j4] = 255 as u8;
-                    }
-                    x = x + x_step;
-                }
-            };
-            if threaded {
-                result.par_chunks_mut(4 * w).enumerate().for_each(process);
-            } else {
-                result.chunks_exact_mut(4 * w).enumerate().for_each(process);
-            }
-
-            //   println!("img elapsed:{}", now.elapsed().as_millis());
-            result
+            Pic::Mono(data) => data.get_rgba8::<S>(threaded, pics, w, h, t),
+            Pic::Grayscale(data) => data.get_rgba8::<S>(threaded, pics, w, h, t),
+            Pic::Gradient(data) => data.get_rgba8::<S>(threaded, pics, w, h, t),
+            Pic::RGB(data) => data.get_rgba8::<S>(threaded, pics, w, h, t),
+            Pic::HSV(data) => data.get_rgba8::<S>(threaded, pics, w, h, t),
         }
     }
 }
@@ -824,70 +408,6 @@ pub fn extract_line_number(token: &Token) -> usize {
     match token {
         Token::OpenParen(ln) | Token::CloseParen(ln) => *ln,
         Token::Constant(_, ln) | Token::Operation(_, ln) => *ln,
-    }
-}
-
-#[inline(always)]
-fn wrap_0_1<S: Simd>(v: S::Vf32) -> S::Vf32 {
-    unsafe {
-        let mut r = S::setzero_ps();
-        for i in 0..S::VF32_WIDTH {
-            r[i] = v[i] % 1.0001;
-        }
-        r
-    }
-}
-
-fn hsv_to_rgb<S: Simd>(h: S::Vf32, s: S::Vf32, v: S::Vf32) -> (S::Vf32, S::Vf32, S::Vf32) {
-    unsafe {
-        let six = S::set1_ps(6.0);
-        let one = S::set1_ps(1.0);
-        let hi = S::cvtps_epi32(S::fast_floor_ps(h * six));
-        let f = h * six - S::cvtepi32_ps(hi);
-        let p = v * (one - s);
-        let q = v * (one - f * s);
-        let t = v * (one - (one - f) * s);
-
-        let mut r = S::setzero_ps();
-        let mut g = S::setzero_ps();
-        let mut b = S::setzero_ps();
-
-        for i in 0..S::VF32_WIDTH {
-            match hi[i] % 6 {
-                0 => {
-                    r[i] = v[i];
-                    g[i] = t[i];
-                    b[i] = p[i];
-                }
-                1 => {
-                    r[i] = q[i];
-                    g[i] = v[i];
-                    b[i] = p[i];
-                }
-                2 => {
-                    r[i] = p[i];
-                    g[i] = v[i];
-                    b[i] = t[i];
-                }
-                3 => {
-                    r[i] = p[i];
-                    g[i] = q[i];
-                    b[i] = v[i];
-                }
-                4 => {
-                    r[i] = t[i];
-                    g[i] = p[i];
-                    b[i] = v[i];
-                }
-                _ => {
-                    r[i] = v[i];
-                    g[i] = p[i];
-                    b[i] = q[i];
-                }
-            }
-        }
-
-        (r, g, b)
     }
 }
